change util to server, add host

+++ src/app/api/users/[id]/route.ts
@@ -8,20 +8,27 @@ export async function GET(_: Request, { params }: { params: { id: string } }) {
     return Response.json(user, { status: 200 });
 }
 
-export async function PATCH(req: Request, { params }: { params: { id: string } }) {
+export async function PATCH(
+    req: Request,
+    { params }: { params: { id: string } },
+) {
     const reqBody = await req.json();
-    const user = await User.findByIdAndUpdate(
-        { _id: params.id },
-        reqBody,
-        { new: true }
-    )
+    const user = await User.findByIdAndUpdate({ _id: params.id }, reqBody, {
+        new: true,
+    });
     if (reqBody.password) {
         user?.save();
     }
-    return Response.json(user, { status: 200 })
+    return Response.json(user, { status: 200 });
 }
 
-export async function DELETE(_: Request, { params }: { params: { id: string } }) {
+export async function DELETE(
+    req: Request,
+    { params }: { params: { id: string } },
+) {
+    if (!req.headers.get("isAdmin")) {
+        return Response.json({ error: "Unauthorized" }, { status: 401 });
+    }
     await User.findByIdAndDelete(params.id);
     return Response.json({}, { status: 200 });
 }
